(BSR) rewrite createLabels
