+++ src/features/bookings/components/BookingsCountV2.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import { Animated } from 'react-native'
 import styled from 'styled-components/native'
 
+import { createBookingsLabels } from 'features/bookings/helpers/createBookingsLabels'
 import { useScaleAnimation } from 'features/favorites/hooks/useScaleFavoritesAnimation'
-import { createLabels } from 'shared/handleTooManyCount/countUtils'
 import { Bookings } from 'ui/svg/icons/Bookings'
 import { AccessibleIcon } from 'ui/svg/icons/types'
 import { Typo } from 'ui/theme'
@@ -15,7 +15,7 @@ export const BookingsCountV2: React.FC<AccessibleIcon> = ({ size, color, testID,
     return <Bookings size={size} color={color} testID={testID} />
   }
 
-  const { fullCountLabel, accessibilityLabel } = createLabels(badgeValue, 'réservations')
+  const { fullCountLabel, accessibilityLabel } = createBookingsLabels(badgeValue)
 
   return (
     <Container>
